feat: create action addproperty and use it instead of calling api route
