+++ app/api/properties/route.js
@@ -1,9 +1,10 @@
-import connectDB from '@/config/database'
-import Property from '@/models/Property'
-import { getSessionUser } from '@/utils/getSessionUser'
-import cloudinary from '@/config/cloudinary'
+// import connectDB from '@/config/database'
+// import Property from '@/models/Property'
+// import { getSessionUser } from '@/utils/getSessionUser'
+// import cloudinary from '@/config/cloudinary'
 
 // TO_REMOVE (not used anymore in the project, but keeping it for reference)
+
 // GET /api/properties
 // export const GET = async (request) => {
 //   try {
@@ -33,87 +34,87 @@ import cloudinary from '@/config/cloudinary'
 // }
 
 // POST /api/properties
-export const POST = async (request) => {
-  try {
-    await connectDB()
+// export const POST = async (request) => {
+//   try {
+//     await connectDB()
 
-    const sessionUser = await getSessionUser()
+//     const sessionUser = await getSessionUser()
 
-    if (!sessionUser || !sessionUser.userId) {
-      return new Response('User ID is required', { status: 401 })
-    }
+//     if (!sessionUser || !sessionUser.userId) {
+//       return new Response('User ID is required', { status: 401 })
+//     }
 
-    const { userId } = sessionUser
+//     const { userId } = sessionUser
 
-    const formData = await request.formData()
+//     const formData = await request.formData()
 
-    // access all values from amenities and images
-    const amenities = formData.getAll('amenities')
-    const images = formData
-      .getAll('images')
-      .filter((image) => image.name !== '')
+//     // access all values from amenities and images
+//     const amenities = formData.getAll('amenities')
+//     const images = formData
+//       .getAll('images')
+//       .filter((image) => image.name !== '')
 
-    // create propertyData object for db
-    const propertyData = {
-      type: formData.get('type'),
-      name: formData.get('name'),
-      description: formData.get('description'),
-      location: {
-        street: formData.get('location.street'),
-        city: formData.get('location.city'),
-        state: formData.get('location.state'),
-        zipcode: formData.get('location.zipcode'),
-      },
-      beds: formData.get('beds'),
-      baths: formData.get('baths'),
-      square_feet: formData.get('square_feet'),
-      amenities,
-      rates: {
-        weekly: formData.get('rates.weekly'),
-        monthly: formData.get('rates.monthly'),
-        nightly: formData.get('rates.nightly'),
-      },
-      seller_info: {
-        name: formData.get('seller_info.name'),
-        email: formData.get('seller_info.email'),
-        phone: formData.get('seller_info.phone'),
-      },
-      owner: userId,
-    }
+//     // create propertyData object for db
+//     const propertyData = {
+//       type: formData.get('type'),
+//       name: formData.get('name'),
+//       description: formData.get('description'),
+//       location: {
+//         street: formData.get('location.street'),
+//         city: formData.get('location.city'),
+//         state: formData.get('location.state'),
+//         zipcode: formData.get('location.zipcode'),
+//       },
+//       beds: formData.get('beds'),
+//       baths: formData.get('baths'),
+//       square_feet: formData.get('square_feet'),
+//       amenities,
+//       rates: {
+//         weekly: formData.get('rates.weekly'),
+//         monthly: formData.get('rates.monthly'),
+//         nightly: formData.get('rates.nightly'),
+//       },
+//       seller_info: {
+//         name: formData.get('seller_info.name'),
+//         email: formData.get('seller_info.email'),
+//         phone: formData.get('seller_info.phone'),
+//       },
+//       owner: userId,
+//     }
 
-    // upload image(s) to cloudinary
-    const imageUploadPromises = []
-    for (const image of images) {
-      const imageBuffer = await image.arrayBuffer()
-      const imageArray = Array.from(new Uint8Array(imageBuffer))
-      const imageData = Buffer.from(imageArray)
-      // convert the image data to base64
-      const imageBase64 = imageData.toString('base64')
-      // make request to upload to cloudinary
-      const result = await cloudinary.uploader.upload(
-        `data:image/png;base64,${imageBase64}`,
-        {
-          folder: 'propertypulse',
-        }
-      )
-      imageUploadPromises.push(result.secure_url)
-      // wait for all images to upload
-      const uploadedImages = await Promise.all(imageUploadPromises)
-      // add uploaded images to propertyData object
-      propertyData.images = uploadedImages
-    }
+//     // upload image(s) to cloudinary
+//     const imageUploadPromises = []
+//     for (const image of images) {
+//       const imageBuffer = await image.arrayBuffer()
+//       const imageArray = Array.from(new Uint8Array(imageBuffer))
+//       const imageData = Buffer.from(imageArray)
+//       // convert the image data to base64
+//       const imageBase64 = imageData.toString('base64')
+//       // make request to upload to cloudinary
+//       const result = await cloudinary.uploader.upload(
+//         `data:image/png;base64,${imageBase64}`,
+//         {
+//           folder: 'propertypulse',
+//         }
+//       )
+//       imageUploadPromises.push(result.secure_url)
+//       // wait for all images to upload
+//       const uploadedImages = await Promise.all(imageUploadPromises)
+//       // add uploaded images to propertyData object
+//       propertyData.images = uploadedImages
+//     }
 
-    const newProperty = new Property(propertyData)
-    await newProperty.save()
+//     const newProperty = new Property(propertyData)
+//     await newProperty.save()
 
-    return Response.redirect(
-      `${process.env.NEXTAUTH_URL}/properties/${newProperty._id}`
-    )
+//     return Response.redirect(
+//       `${process.env.NEXTAUTH_URL}/properties/${newProperty._id}`
+//     )
 
-    // return new Response(JSON.stringify({ message: 'Success' }), {
-    //   status: 201,
-    // })
-  } catch (error) {
-    return new Response('Failed to add property', { status: 500 })
-  }
-}
+//     // return new Response(JSON.stringify({ message: 'Success' }), {
+//     //   status: 201,
+//     // })
+//   } catch (error) {
+//     return new Response('Failed to add property', { status: 500 })
+//   }
+// }
